+++ contracts/multisig/contracts/MultiSigGuardianWalletUpgradeable.sol
@@ -41,13 +41,15 @@ contract MultiSigGuardianWalletUpgradeable is Initializable, UUPSUpgradeable, Mu
      *
      * - The array of wallet owners must not be empty.
      * - The number of required approvals must not be zero and must not exceed the length of the wallet owners array.
-     * - All guardians must be in the wallet owners list.
-     * - The number of required guardian approvals must not be zero and must not exceed the guardians array length.
+     * - If the guardians array is empty, the required guardian approvals must be zero.
+     * - If the guardians array is not empty, all guardians must be in the wallet owners list.
+     * - If the guardians array is not empty, the required guardian approvals must not be zero
+     *   and must not exceed the guardians array length.
      *
      * @param newOwners An array of wallet owners.
      * @param newRequiredApprovals The number of required approvals to execute a transaction.
-     * @param newGuardians An array of guardians (must be a subset of wallet owners).
-     * @param newRequiredGuardianApprovals The number of required guardian approvals to execute a transaction.
+     * @param newGuardians An array of guardians (must be a subset of wallet owners, or empty to disable).
+     * @param newRequiredGuardianApprovals The number of required guardian approvals (zero if guardians disabled).
      */
     function initialize(
         address[] memory newOwners,