+++ contracts/multisig/test/MultiSigWallet.test.ts
@@ -94,7 +94,7 @@ describe("MultiSigWallet contract", () => {
 
   const EXPECTED_VERSION: Version = {
     major: 1,
-    minor: 1,
+    minor: 2,
     patch: 0,
   };