+++ contracts/multisig/test/MultiSigWalletUpgradeable.test.ts
@@ -27,7 +27,7 @@ describe("Contract 'MultiSigWalletUpgradeable'", () => {
 
   const EXPECTED_VERSION: Version = {
     major: 1,
-    minor: 1,
+    minor: 2,
     patch: 0,
   };